refactor: modularize dodo.py into tools/tasks/ package (#152)
Split monolithic dodo.py (1969 lines) into focused modules:
- base.py: shared config and utilities
- install.py: installation tasks
- testing.py: test/coverage tasks
- quality.py: lint/format/type check tasks
- security.py: audit/security tasks
- docs.py: documentation tasks
- git.py: commit/pre-commit tasks
- build.py: build/publish tasks
- maintenance.py: cleanup/update tasks
- release.py: release workflow tasks
- github.py: issue/PR creation tasks

dodo.py reduced to 81-line import file.
All 33 tasks remain functional.
