+++ tools/__init__.py
@@ -1,0 +1,1 @@
+"""Tools package for the pyproject-template."""
