+++ tools/tasks/__init__.py
@@ -1,0 +1,4 @@
+"""Dodo task modules for the pyproject-template.
+
+This package contains modular doit task definitions organized by functionality.
+"""
